Create utterances.json (#47)
Hello Samuel, @Samuel-Jessee as discussed in the email chain from Jackie Ferguson. We are starting to look to onboard CF - Public into Chatbot. In order for Chatbot to 'know' about the FAQs and troubleshooting questions we have created utterances to support these questions. If you could please approve and merge this utterance file we would appreciate it, following it being pushed to production we will then build the intents in Chatbot. Please reach out with any questions, Claire